hw1.12/requirements.txt
asyncio==3.4.3
aiofiles==22.1.0
aiohttp==3.8.3
pyYAML==6.0
pathlib==1.0.1
